pyproject.toml
[project]
name = "ospm"
version = "0.1.0"
description = "Open Source password manager"
authors = [
    {name = "Anvarys",email = "105349498+Anvarys@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click (>=8.3.1,<9.0.0)",
    "clipboard (>=0.0.4,<0.0.5)",
    "pynacl (>=1.6.1,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "platformdirs (>=4.5.1,<5.0.0)"
]
packages = [{include = "ospm"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
ospm = "ospm.__main__:main"